refactor(date utils): tidy up test structure and descriptions
